ported csv download button to svelte, removed unnecessary lit components from lit bundle
