+++ consultation_analyser/lit/csr/index.mjs
@@ -1,9 +1,8 @@
-// import IaiLitCsrExample from "./IaiLitCsrExample/iai-lit-csr-example.lit.csr.mjs";
 import IaiExpandingText from "./IaiExpandingText/iai-expanding-text.lit.csr.mjs";
 import IaiTextWithFallback from "./IaiTextWithFallback/iai-text-with-fallback.lit.csr.mjs";
-// import IaiDataTable from "./IaiDataTable/iai-data-table.lit.csr.mjs";
-import IaiCsvDownload from "./IaiCsvDownload/iai-csv-download.lit.csr.mjs";
-// import IaiRadioInput from "./inputs/IaiRadioInput/iai-radio-input.lit.csr.mjs";
+// import IaiCsvDownload from "./IaiCsvDownload/iai-csv-download.lit.csr.mjs";
+// import Button from "./silverDashboard/Button/button.lit.csr.mjs";
+// import IaiIcon from "./IaiIcon/iai-icon.mjs";
 
 // import IaiResponseFilters from "./IaiResponseFilters/iai-response-filters.lit.csr.mjs";
 // import IaiNumberInput from "./inputs/IaiNumberInput/iai-number-input.lit.csr.mjs";
@@ -16,25 +15,24 @@ import IaiCsvDownload from "./IaiCsvDownload/iai-csv-download.lit.csr.mjs";
 // import IaiQuestionOverview from "./questionsArchive/IaiQuestionOverview/iai-question-overview.lit.csr.mjs";
 // import IaiQuestionTopbar from "./questionsArchive/IaiQuestionTopbar/iai-question-topbar.lit.csr.mjs";
 // import IaiQuestionBody from "./questionsArchive/IaiQuestionBody/iai-question-body.lit.csr.mjs";
-import IaiIcon from "./IaiIcon/iai-icon.mjs";
-import IaiIconButton from "./questionsArchive/IaiIconButton/iai-icon-button.lit.csr.mjs";
+// import IaiIconButton from "./questionsArchive/IaiIconButton/iai-icon-button.lit.csr.mjs";
 // import IaiQuestionOverviewSubtitle from "./questionsArchive/IaiQuestionOverviewSubtitle/iai-question-overview-subtitle.lit.csr.mjs";
 // import IaiMultiResponseItem from "./questionsArchive/IaiMultiResponseItem/iai-multi-response-item.lit.csr.mjs";
 // import IaiTextResponseItem from "./questionsArchive/IaiTextResponseItem/iai-text-response-item.lit.csr.mjs";
-import IaiPageTitle from "./IaiPageTitle/iai-page-title.lit.csr.mjs";
+// import IaiPageTitle from "./IaiPageTitle/iai-page-title.lit.csr.mjs";
 // import IaiExpandingPill from "./IaiExpandingPill/iai-expanding-pill.lit.csr.mjs";
-// import IaiChip from "./IaiChip/iai-chip.lit.csr.mjs";
+// import IaiChip from "./IaiChip/iai-chip.lit.csr.mjs"
 // import IaiResponseDashboard from "./IaiResponseDashboard/iai-response-dashboard.lit.csr.mjs";
-import IaiProgressBar from "./IaiProgressBar/iai-progress-bar.lit.csr.mjs";
-import ProgressBar from "./silverDashboard/ProgressBar/progress-bar.lit.csr.mjs";
-import ProgressCard from "./silverDashboard/questionDetail/ProgressCard/progress-card.lit.csr.mjs";
-import ProgressCards from "./silverDashboard/questionDetail/ProgressCards/progress-cards.lit.csr.mjs";
-import IaiAnimatedNumber from "./IaiAnimatedNumber/iai-animated-number.lit.csr.mjs";
+// import IaiProgressBar from "./IaiProgressBar/iai-progress-bar.lit.csr.mjs";
+// import ProgressBar from "./silverDashboard/ProgressBar/progress-bar.lit.csr.mjs";
+// import ProgressCard from "./silverDashboard/questionDetail/ProgressCard/progress-card.lit.csr.mjs";
+// import ProgressCards from "./silverDashboard/questionDetail/ProgressCards/progress-cards.lit.csr.mjs";
+// import IaiAnimatedNumber from "./IaiAnimatedNumber/iai-animated-number.lit.csr.mjs";
 // import IaiVirtualList from "./IaiVirtualList/iai-virtual-list.lit.csr.mjs";
-import IaiLoadingIndicator from "./IaiLoadingIndicator/iai-loading-indicator.lit.csr.mjs";
+// import IaiLoadingIndicator from "./IaiLoadingIndicator/iai-loading-indicator.lit.csr.mjs";
 // import QuestionOverviewPage from "./silverDashboard/questionOverview/QuestionOverviewPage/question-overview-page.lit.csr.mjs";
 // import SearchBox from "./silverDashboard/inputs/SearchBox/search-box.lit.csr.mjs";
-import Title from "./silverDashboard/Title/title.lit.csr.mjs";
+// import Title from "./silverDashboard/Title/title.lit.csr.mjs";
 // import Panel from "./silverDashboard/Panel/panel.lit.csr.mjs";
 // import CrossSearchCard from "./silverDashboard/CrossSearchCard/cross-search-card.lit.csr.mjs";
 // import Tag from "./silverDashboard/Tag/tag.lit.csr.mjs";
