+++ legacy-frontend/lit/index.js
@@ -372,1029 +372,4 @@ class IaiTextWithFallback extends IaiLitBase {
     }
 }
 customElements.define("iai-text-with-fallback", IaiTextWithFallback);
-
-class IaiIcon extends IaiLitBase {
-    static properties = {
-        ...IaiLitBase.properties,
-        name: { type: String },
-        color: { type: String },
-        fill: {type: Number }, //  0 | 1
-        opsz: { type: Number },
-        wght: { type: Number },
-    }
-
-    static styles = [
-        IaiLitBase.styles,
-        i$4`
-            iai-icon {
-                display: flex;
-            }
-        `
-    ]
-
-    constructor() {
-        super();
-        this.contentId = this.generateId();
-
-        // Google expect icon names to be alphabetically sorted
-        this._ALL_ICON_NAMES = ["visibility", "close", "star", "search", "thumb_up", "thumb_down", "thumbs_up_down", "arrow_drop_down_circle", "download", "diamond", "progress_activity", "sort", "schedule", "calendar_month", "group", "description", "monitoring", "settings", "help", "chat_bubble", "wand_stars", "lan", "finance", "filter_alt", "network_intelligence", "arrow_downward", "arrow_upward", "report", "chevron_left", "keyboard_arrow_down"];
-        this._URL = "https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200&icon_names=" + [...this._ALL_ICON_NAMES].sort().join(",");
-
-        // Prop defaults
-        this.name = "";
-        this.color = "";
-        this.fill = 0;
-        this.opsz = 48;
-        this.wght = 300;
-        
-        this.applyStaticStyles("iai-icon", IaiIcon.styles);
-    }
-
-    firstUpdated() {
-        this.addIconImport();
-    }
-
-    addIconImport() {
-        // Do not add if already added
-        if (document.querySelector(`link[href="${this._URL}"]`)) {
-            return;
-        }
-
-        const linkElement = document.createElement("link");
-        linkElement.rel = "stylesheet";
-        linkElement.href = this._URL;
-        document.head.append(linkElement);
-    }
-
-    render() {
-        return x`
-            <style>
-                #${this.contentId}.material-symbols-outlined {
-                    font-variation-settings:
-                        'FILL' ${this.fill},
-                        'opsz' ${this.opsz},
-                        'wght' ${this.wght};
-                    color: ${this.color};
-                }
-            </style>
-            <span id=${this.contentId} class="material-symbols-outlined">
-                ${this.name}
-            </span>
-        `;
-    }
-}
-customElements.define("iai-icon", IaiIcon);
-
-class Button extends IaiLitBase {
-    static properties = {
-        ...IaiLitBase.properties,
-        text: { type: String },
-        handleClick: { type: Function },
-    }
-
-    static styles = [
-        IaiLitBase.styles,
-        i$4`
-            iai-silver-button button {
-                background: white;
-                outline: none;
-                border: 1px solid var(--iai-silver-color-mid-light);
-                padding: 0.5em 1em;
-                border-radius: 0.5em;
-                cursor: pointer;
-                transition: background 0.3s ease-in-out;
-            }
-            iai-silver-button button:hover {
-                background: var(--iai-silver-color-light);
-            }
-            iai-silver-button button:focus-visible {
-                outline: 3px solid #fd0;
-                outline-offset: 0;
-                box-shadow: inset 0 0 0 2px;
-            }
-        `
-    ]
-
-    constructor() {
-        super();
-        this.contentId = this.generateId();
-
-        // Prop defaults
-        this.text = "white";
-        this.handleClick = () => {};
-
-        this.applyStaticStyles("iai-silver-button", Button.styles);
-    }
-
-    render() {
-        return x`
-            <button @click=${this.handleClick}>
-                ${this.text}
-            </button>
-        `
-    }
-}
-customElements.define("iai-silver-button", Button);
-
-class IaiCsvDownload extends IaiLitBase {
-    static styles = [
-        IaiLitBase.styles,
-        i$4`
-            iai-csv-download a {
-                text-decoration: none;
-            }
-            iai-csv-download a.govuk-button {
-                min-height: auto;
-                min-width: 13em;
-                justify-content: center;
-            }
-            iai-csv-download iai-silver-button button {
-                padding-block: 0.25em;
-            }
-        `
-    ]
-
-    static properties = {
-        ...IaiLitBase.properties,
-        data: {type: Array},
-        fileName: { type: String },
-        variant: { type: String }, // "" | "silver"
-    }
-
-    constructor() {
-        super();
-
-        this.data = [];
-        this.fileName = "data.csv";
-        this.variant = "";
-
-        this.applyStaticStyles("iai-csv-download", IaiCsvDownload.styles);
-    }
-
-    buildCsv(data) {
-        if (!data || !Object.keys(data).length > 0) {
-            return "";
-        }
-
-        const localData = Array.isArray(data) ? data : [data];
-        
-        const keys = Object.keys(data[0]);
-        const rows = [
-            keys.join(","),
-            ...localData.map(row => keys.map(key => JSON.stringify(row[key] ?? "")).join(","))
-        ];
-        return rows.join("\n");
-    }
-
-    getDownloadUrl = (csvContent) => {
-        return "data:text/csv;base64," + btoa(csvContent);
-    }
-
-    render() {
-        return x`
-            <a
-                class=${!this.variant ? "govuk-button" : ""}
-                aria-label="Download themes as CSV"
-                title="Download themes as CSV"
-                href=${this.getDownloadUrl(this.buildCsv(this.props.data || this.data))}
-                download=${this.fileName}
-            >
-                ${this.variant === "silver"
-                    ? x`
-                        <iai-silver-button
-                            .icon=${"download"}
-                            .text=${x`
-                                <iai-icon
-                                    name="download"
-                                ></iai-icon>
-                                <span>
-                                    Export
-                                </span>
-                            `}
-                        ></iai-silver-button>
-                    `
-                    : x`
-                        Download CSV
-                        <iai-icon
-                            name="download"
-                        ></iai-icon>
-                    `}
-            </a>
-        `
-    }
-}
-customElements.define("iai-csv-download", IaiCsvDownload);
-
-class IaiIconButton extends IaiLitBase {
-    static properties = {
-        ...IaiLitBase.properties,
-        title: { type: String },
-        handleClick: { type: Function },
-    }
-
-    static styles = [
-        IaiLitBase.styles,
-        i$4`
-            iai-icon-button button {
-                background: none;
-                border: none;
-                cursor: pointer;
-                border-radius: 0.5em;
-                padding: 0.3em 0.5em;
-                transition: 0.3s ease-in-out;
-                transition-property: background-color, color;
-            }
-            iai-icon-button button:hover {
-                color: black;
-                background: #cbfbf1;
-            }
-            iai-icon-button iai-icon {
-                font-size: 1.2em;
-            }
-            iai-icon-button iai-icon iai-icon {
-                position: absolute;
-            }
-        `
-    ]
-
-    constructor() {
-        super();
-        this.contentId = this.generateId();
-
-        this._SLOT_NAMES = ["icon"];
-
-        // Prop defaults
-        this.title = "";
-        this.handleClick = () => {};
-        
-        this.applyStaticStyles("iai-icon-button", IaiIconButton.styles);
-    }
-
-    updated() {
-        this._SLOT_NAMES.forEach(slotName => this.applySlots(slotName));
-    }
-
-    render() {
-        return x`
-            <button
-                type="button"
-                title=${this.title}
-                @click=${this.handleClick}
-            >
-                <slot name="icon"></slot>
-            </button>
-        `;
-    }
-}
-customElements.define("iai-icon-button", IaiIconButton);
-
-class IaiPageTitle extends IaiLitBase {
-    static properties = {
-        ...IaiLitBase.properties,
-        title: { type: String },
-        subtitle: { type: String },
-    }
-
-    static styles = [
-        IaiLitBase.styles,
-        i$4``
-    ]
-
-    constructor() {
-        super();
-        this.contentId = this.generateId();
-
-        // Prop defaults
-        this.title = "";
-        this.subtitle = "";
-        
-        this.applyStaticStyles("iai-page-title", IaiPageTitle.styles);
-    }
-
-    render() {
-        return x`
-            <div class="govuk-grid-row govuk-!-margin-top-5">
-                <div class="govuk-grid-column-three-quarters-from-desktop tile-panel">
-                    <h2 class="govuk-heading-s">${this.subtitle}</h2>
-                    <h1 class="govuk-heading-l">${this.title}</h1>
-                </div>
-            </div>
-        `;
-    }
-}
-customElements.define("iai-page-title", IaiPageTitle);
-
-class IaiProgressBar extends IaiLitBase {
-    static properties = {
-        ...IaiLitBase.properties,
-        value: { type: Number },
-        label: { type: String },
-    }
-
-    static styles = [
-        IaiLitBase.styles,
-        i$4`
-            iai-progress-bar .container {
-                border: 1px dotted;
-            }
-            iai-progress-bar .container .bar {
-                display: flex;
-                justify-content: end;
-                align-items: center;
-                position: relative;
-                height: 2em;
-                color: white;
-                transition: width 1s ease-in-out;
-                background: var(--iai-colour-pink);
-            }
-            iai-progress-bar .container .label {
-                display: block;    
-                position: absolute;
-                right: 0.5em;
-                text-align: right;
-                color: white;
-                font-weight: bold;
-            }
-            iai-progress-bar .container.low-value .label {
-                left: calc(100% + 0.5em);
-                color: var(--iai-colour-pink);
-                font-size: 1.2em;
-            }
-        `
-    ]
-
-    constructor() {
-        super();
-        this.contentId = this.generateId();
-
-        // Prop defaults
-        this.value = 0;
-        this.label = "";
-        
-        this.applyStaticStyles("iai-progress-bar", IaiProgressBar.styles);
-    }
-
-    getWidth() {
-        if (this.value > 100) {
-            return 100;
-        }
-        if (this.value < 0) {
-            return 0;
-        }
-        return this.value;
-    }
-
-    render() {
-        return x`
-            <div class=${"container" + (this.value < 30 ? " low-value" : "")}>
-                <div class=${"bar" + (this.value >= 100 ? " full" : "")} style=${`width: ${this.getWidth()}%;`}>
-                    ${this.label
-                        ? x`
-                            <span class="label">
-                                ${this.label}
-                            </span>
-                        `
-                        : ""
-                    }
-                </div>
-            </div>
-        `;
-    }
-}
-customElements.define("iai-progress-bar", IaiProgressBar);
-
-class ProgressBar extends IaiLitBase {
-    static properties = {
-        ...IaiLitBase.properties,
-        value: { type: Number },
-        variant: { type: String }, //  "primary" | "secondary"
-    }
-
-    static styles = [
-        IaiLitBase.styles,
-        i$4`
-            iai-silver-progress-bar {
-                flex-grow: 1;
-            }
-            iai-silver-progress-bar iai-progress-bar .container {
-                border-radius: 0.5em;
-                background: var(--iai-silver-color-mid);
-                border: none;
-                opacity: 0.7;
-            }
-            iai-silver-progress-bar .bar {
-                height: 0.5em;
-                border-top-left-radius: 0.5em;
-                border-bottom-left-radius: 0.5em;
-                background: var(--iai-silver-color-dark);
-            }
-            iai-silver-progress-bar iai-progress-bar.primary .bar {
-                background: var(--iai-silver-color-accent);
-            }
-            iai-silver-progress-bar iai-progress-bar.primary .container {
-                background: var(--iai-silver-color-mid-light);
-            }
-        `
-    ]
-
-    constructor() {
-        super();
-        this.contentId = this.generateId();
-
-        this.value = 0;
-        this.variant = "primary";
-        
-        this.applyStaticStyles("iai-silver-progress-bar", ProgressBar.styles);
-    }
-
-    render() {
-        return x`
-            <iai-progress-bar
-                class=${this.variant === "primary" ? "primary" : "secondary"}
-                .value=${this.value}
-                .label=${""}
-            ></iai-progress-bar>
-        `;
-    }
-}
-customElements.define("iai-silver-progress-bar", ProgressBar);
-
-class IconTile extends IaiLitBase {
-    static properties = {
-        ...IaiLitBase.properties,
-        color: { type: String },
-        backgroundColor: { type: String },
-        name: { type: String },
-    }
-
-    static styles = [
-        IaiLitBase.styles,
-        i$4`
-            iai-silver-icon-tile {
-                display: block;    
-                width: max-content;
-            }
-            iai-silver-icon-tile .icon-tile-container {
-                display: flex;    
-                justify-content: center;
-                align-items: center;
-                background: salmon;
-                padding: 0.4em;
-                border-radius: 0.5em;
-            }
-            iai-silver-icon-tile .material-symbols-outlined {
-                font-size: 2em;
-            }
-        `
-    ]
-
-    constructor() {
-        super();
-        this.contentId = this.generateId();
-
-        // Prop defaults
-        this.color = "white";
-        this.backgroundColor = "black";
-        this.name = "";
-
-        this.applyStaticStyles("iai-silver-icon-tile", IconTile.styles);
-    }
-
-    render() {
-        return x`
-            <style>
-                #${this.contentId} {
-                    color: ${this.color};
-                    background: ${this.backgroundColor};
-                }
-            </style>
-            <div id=${this.contentId} class="icon-tile-container">
-                <iai-icon
-                    .name=${this.name}
-                ></iai-icon>
-            </div>
-        `
-    }
-}
-customElements.define("iai-silver-icon-tile", IconTile);
-
-class Title extends IaiLitBase {
-    static properties = {
-        ...IaiLitBase.properties,
-        level: { type: Number },
-        text: { type: String },
-        subtext: { type: String },
-        icon: { type: String },
-        aside: { type: String },
-    }
-
-    static styles = [
-        IaiLitBase.styles,
-        i$4`
-            iai-silver-title h1,
-            iai-silver-title h2,
-            iai-silver-title h3,
-            iai-silver-title h4 {
-                margin: 0;
-            }
-            iai-silver-title h1 {
-                font-size: 1.3em;
-            }
-            iai-silver-title h2 {
-                font-size: 1em;
-            }
-            iai-silver-title h3,
-            iai-silver-title h4 {
-                font-size: 0.9em;
-            }
-            iai-silver-title .container {
-                display: flex;
-                justify-content: space-between;
-                align-items: flex-start;
-            }
-            iai-silver-title .icon-container {
-                display: flex;
-                align-items: center;
-                gap: 0.5em;
-                margin-bottom: 1.5em;
-            }
-        `
-    ]
-
-    constructor() {
-        super();
-        this.contentId = this.generateId();
-
-        // Prop defaults
-        this.level = 2;
-        this.text = "";
-        this.subtext = "";
-        this.icon = "";
-        this.aside = "";
-
-        this.applyStaticStyles("iai-silver-title", Title.styles);
-    }
-
-    renderTitleTag = () => {
-        switch (this.level) {
-            case 1:
-                return x`<h1>${this.text}</h1>`;
-            case 2:
-                return x`<h2>${this.text}</h2>`;
-            case 3:
-                return x`<h3>${this.text}</h3>`;
-            case 4:
-                return x`<h4>${this.text}</h4>`;
-        }
-    }
-
-    render() {
-        return x`
-        <div class="container">
-            <div class="icon-container">
-                ${this.icon
-                    ? x`
-                        <iai-silver-icon-tile
-                            .backgroundColor=${"var(--iai-silver-color-accent-light)"}
-                            .color=${"var(--iai-silver-color-accent)"}
-                            .name=${this.icon}
-                        ></iai-silver-icon-tile>
-                    `
-                    : ""
-                }
-
-                <div>
-                    ${this.renderTitleTag()}
-
-                    ${this.subtext
-                        ? x`
-                            <small>
-                                ${this.subtext}
-                            </small>
-                        `
-                        : ""
-                    }
-                </div>
-            </div>
-
-            <aside>
-                ${this.aside}
-            </aside>
-        </div>
-        `
-    }
-}
-customElements.define("iai-silver-title", Title);
-
-class ProgressCard extends IaiLitBase {
-    static properties = {
-        ...IaiLitBase.properties,
-        title: { type: String },
-        data: { type: Object }, //  key/value pairs
-        _totalCount: { type: Number },
-    }
-
-    static styles = [
-        IaiLitBase.styles,
-        i$4`
-            iai-progress-card article {
-                height: 100%;
-                max-height: 40em;
-                overflow: auto;
-                padding: 1em;
-                font-size: 1em;
-                color: var(--iai-silver-color-text);
-                background: var(--iai-silver-color-light);
-                border-radius: 0.5em;
-            }
-            iai-progress-card ul {
-                display: flex;
-                flex-direction: column;
-                gap: 1.5em;
-                margin: 0;
-                padding-left: 0;
-                font-weight: bold;
-                font-size: 0.9em;
-            }
-            iai-progress-card li {
-                display: flex;
-                justify-content: space-between;
-                align-items: center;
-                flex-direction: column;
-                list-style: none;
-                line-height: 1.5em;
-                gap: 0.5em;
-            }
-            iai-progress-card li>* {
-                width: 100%;
-                word-wrap: break-word;
-            }
-            iai-progress-card li .percentage {
-                font-size: 0.8em;
-                color: rgba(0, 0, 0, 0.5);
-            }
-            iai-progress-card li .count {
-                font-size: 1.2em;
-            }
-            iai-progress-card li {
-                font-size: 0.9em;
-                color: rgba(0, 0, 0, 0.6);
-            }
-            iai-progress-card .info {
-                display: flex;
-                gap: 0.5em;
-            }
-            iai-progress-card .label {
-                font-size: 0.9em;
-            }
-            iai-progress-card .counts {
-                display: flex;
-                align-items: center;
-                gap: 1em;
-                font-weight: normal;
-                font-size: 0.8em;
-            }
-            
-            iai-progress-card iai-silver-progress-bar .container,
-            iai-progress-card iai-silver-progress-bar .container .bar {
-                height: 0.6em;
-            }
-        `
-    ]
-
-    constructor() {
-        super();
-        this.contentId = this.generateId();
-
-        // Prop defaults
-        this.data = {};
-        this._totalCount = 0;
-        
-        this.applyStaticStyles("iai-progress-card", ProgressCard.styles);
-    }
-
-    updated(changedProps) {
-        if (changedProps.has("data")) {
-            this._totalCount = Object.values(this.data).reduce( (a, b) => a + b, 0 );
-        }
-    }
-
-    render() {
-        return x`
-            <article>
-                ${this.title ? x`
-                    <iai-silver-title
-                        .text=${this.title}
-                        .level=${3}
-                    ></iai-silver-title>
-                ` : ""}
-
-                <ul>
-                    ${Object.keys(this.data).map(key => {
-                        const label = key;
-                        const count = this.data[key];
-                        const percentage = this.getPercentage(count, this._totalCount);
-
-                        return x`
-                            <li>
-                                <div class="info">
-                                    <span class="label">
-                                        ${label}
-                                    </span>
-                                    <span class="count">
-                                        ${count.toLocaleString()}
-                                    </span>
-                                    <span class="percentage">
-                                        ${percentage}%
-                                    </span>
-                                </div>
-
-                                <div>
-                                    <iai-silver-progress-bar
-                                        .value=${percentage}
-                                    ></iai-silver-progress-bar>
-                                </div>
-                            </li>
-                        `
-                    })}
-                </ul>
-            </article>
-        `
-    }
-}
-customElements.define("iai-progress-card", ProgressCard);
-
-class Panel extends IaiLitBase {
-    static properties = {
-        ...IaiLitBase.properties,
-        borderColor: { type: String },
-    }
-
-    static styles = [
-        IaiLitBase.styles,
-        i$4`
-            iai-silver-panel .panel {
-                padding: 1em;
-                border: 1px solid;
-                border-radius: 1em;
-            }
-        `
-    ]
-
-    constructor() {
-        super();
-        this.contentId = this.generateId();
-        
-        // Prop defaults
-        this.borderColor = "";
-
-        this.applyStaticStyles("iai-silver-panel", Panel.styles);
-    }
-
-    updated() {
-        this.applySlots("content");
-    }
-
-    render() {
-        return x`
-            <style>
-                #${this.contentId} {
-                    border-color: ${this.borderColor || "var(--iai-silver-color-mid-light)"};
-                }
-            </style>
-            <div id=${this.contentId} class="panel">
-                <slot name="content"></slot>
-            </div>
-        `
-    }
-}
-customElements.define("iai-silver-panel", Panel);
-
-class ProgressCards extends IaiLitBase {
-    static properties = {
-        ...IaiLitBase.properties,
-        data: { type: Array },
-        title: { type: String },
-        subtitle: { type: String },
-        themeFilters: { type: Array },
-        demoFilters: { type: Object},
-        total: { type: Number },
-
-        demoFiltersApplied: { type: Function },
-        themeFiltersApplied: { type: Function },
-    }
-
-    static styles = [
-        IaiLitBase.styles,
-        i$4` 
-            iai-progress-cards .cards {
-                display: flex;
-                flex-wrap: wrap;
-                gap: 1em;
-                max-width: 100%;
-                overflow: auto;
-            }    
-            iai-progress-cards iai-progress-card {
-                flex-grow: 1;
-                max-width: 100%;
-            }
-            .themes-warning .tag {
-                width: 100%;
-                margin-bottom: 1em;
-            }
-        `
-    ]
-
-    constructor() {
-        super();
-        this.contentId = this.generateId();
-        this._MAX_DEMO_ANSWERS = 10;
-
-        // Prop defaults
-        this.title = "";
-        this.subtitle = "";
-        this.data = [];
-        this.themeFilters = [];
-        this.demoFilters = {};
-        this.total = 0;
-
-        this.demoFiltersApplied = () => {};
-        this.themeFiltersApplied = () => {};
-        
-        this.applyStaticStyles("iai-progress-cards", ProgressCards.styles);
-    }
-
-    getFilterWarningText() {
-        const demoFiltersText = Object
-            .values(this.demoFilters)
-            .flat(1)
-            .filter(Boolean)
-            .join(", ");
-
-        const themeFiltersText = this.themeFilters.length
-            ? `${this.themeFilters.length} themes`
-            : "";
-
-        return [demoFiltersText, themeFiltersText].filter(Boolean).join(", ");
-    }
-
-    render() {
-        return x`
-            <iai-silver-panel>
-                <div slot="content">
-                    <div class="top-panel">
-                        <iai-silver-title
-                            .text=${this.title}
-                            .subtext=${this.subtitle}
-                            .icon=${"group"}
-                            .level=${2}
-                        ></iai-silver-title>
-
-                        ${this.themeFiltersApplied() || this.demoFiltersApplied()
-                            ? x`
-                                <iai-silver-tag
-                                    class="themes-warning"
-                                    .text=${"Active theme analysis filters"}
-                                    .subtext=${`Showing data for ${this.total.toLocaleString()} responses (filtered by: ${this.getFilterWarningText()})`}
-                                    .icon=${"report"}
-                                    .status=${"Analysing"}
-                                ></iai-silver-tag>
-                            `
-                            : ""
-                        }
-
-                        <div class="cards">
-                            ${Object.keys(this.data).map(category => {
-                                return Object.values(this.data[category]).length <= this._MAX_DEMO_ANSWERS
-                                    ? x`
-                                        <iai-progress-card
-                                            .title=${this.toTitleCase(category)}
-                                            .data=${this.data[category]}
-                                        ></iai-progress-card>
-                                    ` : ""})}
-                        </div>
-                    </div>
-                </div>
-            </iai-silver-panel>
-        `
-    }
-}
-customElements.define("iai-progress-cards", ProgressCards);
-
-class IaiAnimatedNumber extends IaiLitBase {
-    static properties = {
-        ...IaiLitBase.properties,
-        number: { type: Number },
-        duration: { type: Number }, // miliseconds
-        _displayNumber: { type: Number },
-    }
-
-    static styles = [
-        IaiLitBase.styles,
-        i$4``
-    ]
-
-    constructor() {
-        super();
-        this.contentId = this.generateId();
-
-        // Prop defaults
-        this.number = 0;
-        this.duration = 1000;
-        this._displayNumber = 0;
-        
-        this.applyStaticStyles("iai-animated-number", IaiAnimatedNumber.styles);
-    }
-
-    setDisplayNumber = (newDisplayNumber) => {
-        this._displayNumber = newDisplayNumber;
-    }
-
-    animate = (start, end, duration) => {
-        const setDisplayNumber = (newDisplayNumber) => {
-            this._displayNumber = newDisplayNumber;
-        };
-        const startTime = performance.now();
-
-        function update_number(currTime) {
-            const elapsedTime = currTime - startTime;
-            const time = Math.min(elapsedTime / duration, 1);
-            const currValue = start * (1 - time) + end * time;
-
-            // round to 1 decimal point
-            setDisplayNumber(Math.round(currValue * 10) / 10);
-
-            if (time < 1) {
-                requestAnimationFrame(update_number);
-            }
-        }
-
-        requestAnimationFrame(update_number);
-    }
-
-    updated(changedProps) {
-        if (changedProps.has("number")) {
-            this.animate(this._displayNumber, this.number, this.duration);
-        }
-    }
-
-    render() {
-        return x`
-            <span>${this._displayNumber}</span>
-        `;
-    }
-}
-customElements.define("iai-animated-number", IaiAnimatedNumber);
-
-class IaiLoadingIndicator extends IaiLitBase {
-    static properties = {
-        ...IaiLitBase.properties,
-    }
-
-    static styles = [
-        IaiLitBase.styles,
-        i$4`
-            iai-loading-indicator .spinner {
-                display: flex;
-                justify-content: center;
-                animation-name: spin;
-                animation-duration: 1s;
-                animation-iteration-count: infinite;
-                animation-timing-function: ease-in-out;
-            }
-            iai-loading-indicator .spinner iai-icon .material-symbols-outlined {
-                font-size: 3em;
-            }
-
-            @keyframes spin {
-                from {
-                    transform:rotate(0deg);
-                }
-                to {
-                    transform:rotate(360deg);
-                }
-            }
-        `
-    ]
-
-    constructor() {
-        super();
-        this.contentId = this.generateId();
-        
-        this.applyStaticStyles("iai-loading-indicator", IaiLoadingIndicator.styles);
-    }
-
-    render() {
-        return x`
-            <div class="spinner">
-                <iai-icon
-                    name="progress_activity"
-                    .opsz=${48}
-                    .color=${"var(--iai-colour-pink)"}
-                ></iai-icon>
-            </div>
-        `;
-    }
-}
-customElements.define("iai-loading-indicator", IaiLoadingIndicator);
 //# sourceMappingURL=index.js.map
